requirements.txt
Flask
PyPDF2
pdf2image
Pillow
Werkzeug
